extensions: re-introduce S3ProvisionPipeline
