+++ extensions/aws/s3/provision/src/main/java/org/eclipse/dataspaceconnector/provision/aws/s3/S3BucketProvisioner.java
@@ -50,6 +50,7 @@ import software.amazon.awssdk.services.s3.model.ObjectIdentifier;
 import software.amazon.awssdk.services.sts.StsAsyncClient;
 import software.amazon.awssdk.services.sts.model.AssumeRoleRequest;
 import software.amazon.awssdk.services.sts.model.AssumeRoleResponse;
+import software.amazon.awssdk.services.sts.model.Credentials;
 import software.amazon.awssdk.utils.Pair;
 
 import java.util.concurrent.CompletableFuture;
@@ -62,32 +63,6 @@ import static java.util.stream.Collectors.joining;
  * Asynchronously provisions S3 buckets.
  */
 public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefinition, S3BucketProvisionedResource> {
-
-    // Do not modify this trust policy
-    private static final String ASSUME_ROLE_TRUST = "{" +
-            "  \"Version\": \"2012-10-17\"," +
-            "  \"Statement\": [" +
-            "    {" +
-            "      \"Effect\": \"Allow\"," +
-            "      \"Principal\": {" +
-            "        \"AWS\": \"%s\"" +
-            "      }," +
-            "      \"Action\": \"sts:AssumeRole\"" +
-            "    }" +
-            "  ]" +
-            "}";
-    // Do not modify this bucket policy
-    private static final String BUCKET_POLICY = "{" +
-            "    \"Version\": \"2012-10-17\"," +
-            "    \"Statement\": [" +
-            "        {" +
-            "            \"Sid\": \"TemporaryAccess\", " +
-            "            \"Effect\": \"Allow\"," +
-            "            \"Action\": \"s3:PutObject\"," +
-            "            \"Resource\": \"arn:aws:s3:::%s/*\"" +
-            "        }" +
-            "    ]" +
-            "}";
 
     private final ClientProvider clientProvider;
     private final int sessionDuration;
@@ -127,17 +102,15 @@ public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefiniti
         var iamClient = clientProvider.clientFor(IamAsyncClient.class, resourceDefinition.getRegionId());
         var stsClient = clientProvider.clientFor(StsAsyncClient.class, resourceDefinition.getRegionId());
 
-        var request = CreateBucketRequest.builder()
-                .bucket(resourceDefinition.getBucketName())
-                .createBucketConfiguration(CreateBucketConfiguration.builder().build())
-                .build();
-
-        return s3AsyncClient.createBucket(request)
-                .thenCompose(r -> getUser(iamClient))
-                .thenCompose(response -> createRole(iamClient, resourceDefinition, response))
-                .thenCompose(response -> createRolePolicy(iamClient, resourceDefinition, response))
-                .thenCompose(role -> assumeRole(stsClient, role))
-                .thenApply(result -> provisionSuccedeed(resourceDefinition, result))
+        return S3ProvisionPipeline.Builder.newInstance(retryPolicy)
+                .s3Client(s3AsyncClient)
+                .iamClient(iamClient)
+                .stsClient(stsClient)
+                .sessionDuration(sessionDuration)
+                .monitor(monitor)
+                .build()
+                .provision(resourceDefinition)
+                .thenApply(result -> provisionSuccedeed(resourceDefinition, result.left(), result.right()))
                 .exceptionally(throwable -> provisionFailed(resourceDefinition, throwable));
     }
 
@@ -158,71 +131,17 @@ public class S3BucketProvisioner implements Provisioner<S3BucketResourceDefiniti
                 .thenApply(response -> DeprovisionResponse.Builder.newInstance().ok().resource(resource).build());
     }
 
-    private CompletableFuture<Pair<Role, AssumeRoleResponse>> assumeRole(StsAsyncClient stsClient, Role role) {
-        return Failsafe.with(retryPolicy).getStageAsync(() -> {
-            monitor.debug("S3ProvisionPipeline: attempting to assume the role");
-            AssumeRoleRequest roleRequest = AssumeRoleRequest.builder()
-                    .roleArn(role.arn())
-                    .roleSessionName("transfer")
-                    .externalId("123")
-                    .build();
-
-            return stsClient.assumeRole(roleRequest)
-                    .thenApply(response -> Pair.of(role, response));
-        });
-    }
-
-    private CompletableFuture<Role> createRolePolicy(IamAsyncClient iamAsyncClient, S3BucketResourceDefinition resourceDefinition, CreateRoleResponse response) {
-        return Failsafe.with(retryPolicy).getStageAsync(() -> {
-            Role role = response.role();
-            PutRolePolicyRequest policyRequest = PutRolePolicyRequest.builder()
-                    .policyName(resourceDefinition.getTransferProcessId())
-                    .roleName(role.roleName())
-                    .policyDocument(format(BUCKET_POLICY, resourceDefinition.getBucketName()))
-                    .build();
-
-            monitor.debug("S3ProvisionPipeline: attach bucket policy to role " + role.arn());
-            return iamAsyncClient.putRolePolicy(policyRequest)
-                    .thenApply(policyResponse -> role);
-        });
-    }
-
-    private CompletableFuture<CreateRoleResponse> createRole(IamAsyncClient iamClient, S3BucketResourceDefinition resourceDefinition, GetUserResponse response) {
-        return Failsafe.with(retryPolicy).getStageAsync(() -> {
-            String userArn = response.user().arn();
-            Tag tag = Tag.builder().key("dataspaceconnector:process").value(resourceDefinition.getTransferProcessId()).build();
-
-            monitor.debug("S3ProvisionPipeline: create role for user" + userArn);
-            CreateRoleRequest createRoleRequest = CreateRoleRequest.builder()
-                    .roleName(resourceDefinition.getTransferProcessId()).description("EDC transfer process role")
-                    .assumeRolePolicyDocument(format(ASSUME_ROLE_TRUST, userArn))
-                    .maxSessionDuration(sessionDuration)
-                    .tags(tag)
-                    .build();
-
-            return iamClient.createRole(createRoleRequest);
-        });
-    }
-
-    private CompletableFuture<GetUserResponse> getUser(IamAsyncClient iamAsyncClient) {
-        return Failsafe.with(retryPolicy).getStageAsync(() -> {
-            monitor.debug("S3ProvisionPipeline: get user");
-            return iamAsyncClient.getUser();
-        });
-    }
-
-    private ProvisionResponse provisionSuccedeed(S3BucketResourceDefinition resourceDefinition, Pair<Role, AssumeRoleResponse> result) {
+    private ProvisionResponse provisionSuccedeed(S3BucketResourceDefinition resourceDefinition, Role role, Credentials credentials) {
         monitor.debug("S3ProvisionPipeline: STS credentials obtained, continuing...");
         var resource = S3BucketProvisionedResource.Builder.newInstance()
                 .id(resourceDefinition.getBucketName())
                 .resourceDefinitionId(resourceDefinition.getId())
                 .region(resourceDefinition.getRegionId())
                 .bucketName(resourceDefinition.getBucketName())
-                .role(result.left().roleName())
+                .role(role.roleName())
                 .transferProcessId(resourceDefinition.getTransferProcessId())
                 .build();
 
-        var credentials = result.right().credentials();
         var secretToken = new AwsTemporarySecretToken(credentials.accessKeyId(), credentials.secretAccessKey(), credentials.sessionToken(), credentials.expiration().toEpochMilli());
 
         monitor.debug("Bucket request submitted: " + resourceDefinition.getBucketName());
